Implement true-color support in the terminal

+++ rxvt/rxvtlibrend.h
@@ -3,7 +3,7 @@
 #define ANSI256_AND_TRUECOLOR_defined
 
 #ifdef ANSI256_AND_TRUECOLOR
-typedef unsigned long long rend_t;
+typedef unsigned int rend_t;
 #else
 typedef unsigned int rend_t;
 #endif
